+++ controllers/dependency/dependency_checker.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 
 	"github.com/cloudogu/cesapp-lib/core"
-	"github.com/cloudogu/k8s-registry-lib/dogu/local"
+	"github.com/cloudogu/k8s-registry-lib/dogu"
 )
 
 // DependencyValidator is responsible to validate the dependencies of a dogu
@@ -19,7 +19,7 @@ type CompositeDependencyValidator struct {
 }
 
 // NewCompositeDependencyValidator create a new composite validator checking the dogu and client dependencies
-func NewCompositeDependencyValidator(version *core.Version, doguRegistry local.LocalDoguRegistry) *CompositeDependencyValidator {
+func NewCompositeDependencyValidator(version *core.Version, doguRegistry dogu.LocalRegistry) *CompositeDependencyValidator {
 	validators := []DependencyValidator{}
 
 	operatorDependencyValidator := NewOperatorDependencyValidator(version)