+++ controllers/resource/requirementsGenerator.go
@@ -1,14 +1,15 @@
 package resource
 
 import (
+	"context"
 	"errors"
 	"fmt"
+	"github.com/cloudogu/k8s-registry-lib/config"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	"strings"
 
 	"github.com/cloudogu/cesapp-lib/core"
-	"github.com/cloudogu/cesapp-lib/registry"
 )
 
 type resourceType string
@@ -25,16 +26,19 @@ var resourceTypeMapping = map[resourceType]corev1.ResourceName{
 	storageType: corev1.ResourceEphemeralStorage,
 }
 
-type requirementsGenerator struct {
-	configRegistry registry.Registry
+type RequirementsGenerator struct {
+	doguConfigGetter doguConfigGetter
 }
 
-func NewRequirementsGenerator(configRegistry registry.Registry) *requirementsGenerator {
-	return &requirementsGenerator{configRegistry: configRegistry}
+func NewRequirementsGenerator(doguConfigGetter doguConfigGetter) *RequirementsGenerator {
+	return &RequirementsGenerator{doguConfigGetter: doguConfigGetter}
 }
 
-func (r requirementsGenerator) Generate(dogu *core.Dogu) (corev1.ResourceRequirements, error) {
-	doguConfig := r.configRegistry.DoguConfig(dogu.GetSimpleName())
+func (r RequirementsGenerator) Generate(ctx context.Context, dogu *core.Dogu) (corev1.ResourceRequirements, error) {
+	doguConfig, err := r.doguConfigGetter.Get(ctx, config.SimpleDoguName(dogu.GetSimpleName()))
+	if err != nil {
+		return corev1.ResourceRequirements{}, fmt.Errorf("unable to get config for dogu %s: %w", dogu.GetSimpleName(), err)
+	}
 
 	requirements := corev1.ResourceRequirements{
 		Limits:   corev1.ResourceList{},
@@ -43,7 +47,7 @@ func (r requirementsGenerator) Generate(dogu *core.Dogu) (corev1.ResourceRequire
 
 	var errList []error
 	for resourceType := range resourceTypeMapping {
-		err := appendRequirementsForResourceType(resourceType, requirements, doguConfig, dogu)
+		err := appendRequirementsForResourceType(ctx, resourceType, requirements, doguConfig, dogu)
 		if err != nil {
 			errList = append(errList, err)
 		}
@@ -56,44 +60,39 @@ func (r requirementsGenerator) Generate(dogu *core.Dogu) (corev1.ResourceRequire
 	return requirements, nil
 }
 
-func readFromConfigOrDefault(key string, doguConfig registry.ConfigurationContext, dogu *core.Dogu) (string, error) {
-	configValue, err := doguConfig.Get(key)
-
-	if err != nil {
-		if registry.IsKeyNotFoundError(err) {
-			for _, field := range dogu.Configuration {
-				if field.Name == key {
-					return field.Default, nil
-				}
+func readFromConfigOrDefault(key string, doguConfig config.DoguConfig, dogu *core.Dogu) string {
+	configValue, ok := doguConfig.Get(config.Key(key))
+	if !ok {
+		for _, field := range dogu.Configuration {
+			if field.Name == key {
+				return field.Default
 			}
-
-			return "", nil
 		}
 
-		return "", fmt.Errorf("failed to read value of key '%s' from registry config of dogu '%s': %w", key, dogu.Name, err)
+		return ""
 	}
 
-	return configValue, nil
+	return configValue.String()
 }
 
-func appendRequirementsForResourceType(resourceType resourceType, requirements corev1.ResourceRequirements, doguConfig registry.ConfigurationContext, dogu *core.Dogu) error {
+func appendRequirementsForResourceType(_ context.Context, resourceType resourceType, requirements corev1.ResourceRequirements, doguConfig config.DoguConfig, dogu *core.Dogu) error {
 	resourceName := resourceTypeMapping[resourceType]
 
 	limitKey := fmt.Sprintf("container_config/%s_limit", resourceType)
-	limit, limitErr := readFromConfigOrDefault(limitKey, doguConfig, dogu)
+	limit := readFromConfigOrDefault(limitKey, doguConfig, dogu)
 	var limitConversionErr error
 	if limit != "" {
 		requirements.Limits[resourceName], limitConversionErr = convertCesUnitToQuantity(limit, resourceType)
 	}
 
 	requestKey := fmt.Sprintf("container_config/%s_request", resourceType)
-	request, requestErr := readFromConfigOrDefault(requestKey, doguConfig, dogu)
+	request := readFromConfigOrDefault(requestKey, doguConfig, dogu)
 	var requestConversionErr error
 	if request != "" {
 		requirements.Requests[resourceName], requestConversionErr = convertCesUnitToQuantity(request, resourceType)
 	}
 
-	err := errors.Join(limitErr, limitConversionErr, requestErr, requestConversionErr)
+	err := errors.Join(limitConversionErr, requestConversionErr)
 	if err != nil {
 		return fmt.Errorf("errors occured while appending requirements for resource type '%s': %w", resourceType, err)
 	}