+++ controllers/resource/requirementsUpdater.go
@@ -4,87 +4,67 @@ import (
 	"context"
 	"errors"
 	"fmt"
-
+	k8sv1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
+	"github.com/cloudogu/k8s-registry-lib/config"
 	v1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes"
-
-	"github.com/cloudogu/cesapp-lib/core"
-	"github.com/cloudogu/cesapp-lib/registry"
-	k8sv1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
-	"github.com/cloudogu/k8s-dogu-operator/internal/cloudogu"
-	"github.com/cloudogu/k8s-registry-lib/dogu/local"
-	coreosclient "go.etcd.io/etcd/client/v2"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
-	triggerSyncEtcdKeyFullPath = "/config/_global/sync_resource_requirements"
+	triggerSyncKey = "sync_resource_requirements"
 )
 
-// requirementsUpdater is responsible to update all resource requirements for dogu deployments when a certain trigger is called.
-type requirementsUpdater struct {
-	client            client.Client
-	namespace         string
-	registry          registry.Registry
-	localDoguRegistry local.LocalDoguRegistry
-	requirementsGen   cloudogu.ResourceRequirementsGenerator
+// RequirementsUpdater is responsible to update all resource requirements for dogu deployments when a certain trigger is called.
+type RequirementsUpdater struct {
+	client              client.Client
+	namespace           string
+	globalConfigWatcher globalConfigurationWatcher
+	localDoguRegistry   doguGetter
+	requirementsGen     requirementsGenerator
 }
 
 // NewRequirementsUpdater creates a new runnable responsible to detect changes in the container configuration of dogus.
-func NewRequirementsUpdater(client client.Client, namespace string, clientSet kubernetes.Interface) (*requirementsUpdater, error) {
-	endpoint := fmt.Sprintf("http://etcd.%s.svc.cluster.local:4001", namespace)
-	reg, err := registry.New(core.Registry{
-		Type:      "etcd",
-		Endpoints: []string{endpoint},
-	})
-	if err != nil {
-		return nil, err
-	}
+func NewRequirementsUpdater(client client.Client, namespace string, doguConfigGetter doguConfigGetter, doguReg doguGetter, globalWatcher globalConfigurationWatcher) (*RequirementsUpdater, error) {
+	requirementsGen := NewRequirementsGenerator(doguConfigGetter)
 
-	requirementsGen := NewRequirementsGenerator(reg)
-
-	return &requirementsUpdater{
-		client:    client,
-		namespace: namespace,
-		registry:  reg,
-		localDoguRegistry: local.NewCombinedLocalDoguRegistry(
-			clientSet.CoreV1().ConfigMaps(namespace),
-			reg,
-		),
-		requirementsGen: requirementsGen,
+	return &RequirementsUpdater{
+		client:              client,
+		namespace:           namespace,
+		globalConfigWatcher: globalWatcher,
+		localDoguRegistry:   doguReg,
+		requirementsGen:     requirementsGen,
 	}, nil
 }
 
 // Start is the entry point for the updater.
-func (hlu *requirementsUpdater) Start(ctx context.Context) error {
-	return hlu.startEtcdWatch(ctx)
+func (hlu *RequirementsUpdater) Start(ctx context.Context) error {
+	return hlu.startWatch(ctx)
 }
 
-func (hlu *requirementsUpdater) startEtcdWatch(ctx context.Context) error {
-	ctrl.LoggerFrom(ctx).Info(fmt.Sprintf("Start etcd watcher on certificate key [%s]", triggerSyncEtcdKeyFullPath))
+func (hlu *RequirementsUpdater) startWatch(ctx context.Context) error {
+	ctrl.LoggerFrom(ctx).Info(fmt.Sprintf("Start watching on the trigger for synchronization [%s] in global config", triggerSyncKey))
 
-	triggerChannel := make(chan *coreosclient.Response)
-	go func() {
-		hlu.registry.RootConfig().Watch(ctx, triggerSyncEtcdKeyFullPath, false, triggerChannel)
-	}()
+	watchResChan, err := hlu.globalConfigWatcher.Watch(ctx, config.KeyFilter(triggerSyncKey))
+	if err != nil {
+		return fmt.Errorf("could not start watch for key [%s]: %w", triggerSyncKey, err)
+	}
 
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case <-triggerChannel:
-			err := hlu.triggerSync(ctx)
-			if err != nil {
-				return err
-			}
+	for range watchResChan {
+		lErr := hlu.triggerSync(ctx)
+		if lErr != nil {
+			return lErr
 		}
 	}
+
+	ctrl.LoggerFrom(ctx).Info(fmt.Sprintf("Stopped watching on global config for certificate key [%s] because channel has been closed", triggerSyncKey))
+
+	return nil
 }
 
-func (hlu *requirementsUpdater) triggerSync(ctx context.Context) error {
-	ctrl.LoggerFrom(ctx).Info("Trigger for updating dogu resource requirements detected in registry. Updating deployment for all dogus...")
+func (hlu *RequirementsUpdater) triggerSync(ctx context.Context) error {
+	ctrl.LoggerFrom(ctx).Info("Trigger for updating dogu resource requirements detected in global config. Updating deployment for all dogus...")
 
 	installedDogus, err := hlu.getInstalledDogus(ctx)
 	if err != nil {
@@ -93,31 +73,31 @@ func (hlu *requirementsUpdater) triggerSync(ctx context.Context) error {
 
 	var result error
 	for _, dogu := range installedDogus.Items {
-		doguJson, err := hlu.localDoguRegistry.GetCurrent(ctx, dogu.GetName())
-		if err != nil {
-			result = errors.Join(result, fmt.Errorf("failed to get dogu.json of dogu [%s] from registry: %w", dogu.Name, err))
+		doguJson, lErr := hlu.localDoguRegistry.GetCurrent(ctx, dogu.GetName())
+		if lErr != nil {
+			result = errors.Join(result, fmt.Errorf("failed to get dogu.json of dogu [%s] from registry: %w", dogu.Name, lErr))
 			continue
 		}
 
 		doguIdentifier := types.NamespacedName{Name: dogu.GetName(), Namespace: dogu.GetNamespace()}
 		doguDeployment := &v1.Deployment{}
-		err = hlu.client.Get(ctx, doguIdentifier, doguDeployment)
-		if err != nil {
-			result = errors.Join(result, fmt.Errorf("failed to get deployment of dogu [%s/%s] from cluster: %w", dogu.Namespace, dogu.Name, err))
+		lErr = hlu.client.Get(ctx, doguIdentifier, doguDeployment)
+		if lErr != nil {
+			result = errors.Join(result, fmt.Errorf("failed to get deployment of dogu [%s/%s] from cluster: %w", dogu.Namespace, dogu.Name, lErr))
 			continue
 		}
 
-		requirements, err := hlu.requirementsGen.Generate(doguJson)
-		if err != nil {
-			result = errors.Join(result, fmt.Errorf("failed to generate resource requirements of dogu [%s/%s] in cluster: %w", dogu.Namespace, dogu.Name, err))
+		requirements, lErr := hlu.requirementsGen.Generate(ctx, doguJson)
+		if lErr != nil {
+			result = errors.Join(result, fmt.Errorf("failed to generate resource requirements of dogu [%s/%s] in cluster: %w", dogu.Namespace, dogu.Name, lErr))
 			continue
 		}
 
 		doguDeployment.Spec.Template.Spec.Containers[0].Resources = requirements
 
-		err = hlu.client.Update(ctx, doguDeployment)
-		if err != nil {
-			result = errors.Join(result, fmt.Errorf("failed to update deployment of dogu [%s/%s] in cluster: %w", dogu.Namespace, dogu.Name, err))
+		lErr = hlu.client.Update(ctx, doguDeployment)
+		if lErr != nil {
+			result = errors.Join(result, fmt.Errorf("failed to update deployment of dogu [%s/%s] in cluster: %w", dogu.Namespace, dogu.Name, lErr))
 			continue
 		}
 	}
@@ -125,7 +105,7 @@ func (hlu *requirementsUpdater) triggerSync(ctx context.Context) error {
 	return result
 }
 
-func (hlu *requirementsUpdater) getInstalledDogus(ctx context.Context) (*k8sv1.DoguList, error) {
+func (hlu *RequirementsUpdater) getInstalledDogus(ctx context.Context) (*k8sv1.DoguList, error) {
 	doguList := &k8sv1.DoguList{}
 
 	err := hlu.client.List(ctx, doguList, client.InNamespace(hlu.namespace))