+++ controllers/serviceaccount/serviceaccount_remover.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/cloudogu/k8s-registry-lib/config"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes"
@@ -11,8 +12,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/cloudogu/cesapp-lib/core"
-	"github.com/cloudogu/cesapp-lib/registry"
-	"github.com/cloudogu/k8s-registry-lib/dogu/local"
+	"github.com/cloudogu/k8s-registry-lib/dogu"
 
 	"github.com/cloudogu/k8s-dogu-operator/controllers/exec"
 	"github.com/cloudogu/k8s-dogu-operator/internal/cloudogu"
@@ -21,9 +21,9 @@ import (
 // Remover removes a dogu's service account.
 type remover struct {
 	client            client.Client
-	registry          registry.Registry
+	sensitiveDoguRepo SensitiveDoguConfigRepository
 	doguFetcher       cloudogu.LocalDoguFetcher
-	localDoguRegistry local.LocalDoguRegistry
+	localDoguRegistry dogu.LocalRegistry
 	executor          cloudogu.CommandExecutor
 	clientSet         kubernetes.Interface
 	apiClient         serviceAccountApiClient
@@ -31,10 +31,10 @@ type remover struct {
 }
 
 // NewRemover creates a new instance of ServiceAccountRemover
-func NewRemover(registry registry.Registry, localFetcher cloudogu.LocalDoguFetcher, localDoguRegistry local.LocalDoguRegistry, commandExecutor cloudogu.CommandExecutor, client client.Client, clientSet kubernetes.Interface, namespace string) *remover {
+func NewRemover(repo SensitiveDoguConfigRepository, localFetcher cloudogu.LocalDoguFetcher, localDoguRegistry dogu.LocalRegistry, commandExecutor cloudogu.CommandExecutor, client client.Client, clientSet kubernetes.Interface, namespace string) *remover {
 	return &remover{
 		client:            client,
-		registry:          registry,
+		sensitiveDoguRepo: repo,
 		doguFetcher:       localFetcher,
 		localDoguRegistry: localDoguRegistry,
 		executor:          commandExecutor,
@@ -48,29 +48,26 @@ func NewRemover(registry registry.Registry, localFetcher cloudogu.LocalDoguFetch
 func (r *remover) RemoveAll(ctx context.Context, dogu *core.Dogu) error {
 	logger := log.FromContext(ctx)
 
+	sensitiveConfig, err := r.sensitiveDoguRepo.Get(ctx, config.SimpleDoguName(dogu.GetSimpleName()))
+	if err != nil {
+		return fmt.Errorf("unbale to get sensitive config for dogu %s: %w", dogu.GetSimpleName(), err)
+	}
+
 	var allProblems error
 	for _, serviceAccount := range dogu.ServiceAccounts {
 		switch serviceAccount.Kind {
 		case "":
 			fallthrough
 		case doguKind:
-			err := r.removeDoguServiceAccount(ctx, dogu, serviceAccount)
-			if err != nil {
-				allProblems = errors.Join(allProblems, err)
+			lErr := r.removeDoguServiceAccount(ctx, dogu, serviceAccount, &sensitiveConfig)
+			if lErr != nil {
+				allProblems = errors.Join(allProblems, fmt.Errorf("unable to remove service account for dogu %s: %w", serviceAccount.Type, lErr))
 			}
 		case componentKind:
-			err := r.removeComponentServiceAccount(ctx, dogu, serviceAccount)
-			if err != nil {
-				allProblems = errors.Join(allProblems, err)
+			lErr := r.removeComponentServiceAccount(ctx, dogu, serviceAccount, &sensitiveConfig)
+			if lErr != nil {
+				allProblems = errors.Join(allProblems, fmt.Errorf("unable to remove service account for component %s: %w", serviceAccount.Type, lErr))
 			}
-		case cesKind:
-			if serviceAccount.Type == "cesappd" {
-				err := r.removeCesControlServiceAccount(dogu)
-				if err != nil {
-					allProblems = errors.Join(allProblems, err)
-				}
-			}
-			continue
 		default:
 			logger.Error(fmt.Errorf("unknown service account kind: %s", serviceAccount.Kind), "skipping service account deletion")
 			continue
@@ -80,18 +77,13 @@ func (r *remover) RemoveAll(ctx context.Context, dogu *core.Dogu) error {
 	return allProblems
 }
 
-func (r *remover) removeDoguServiceAccount(ctx context.Context, dogu *core.Dogu, serviceAccount core.ServiceAccount) error {
+func (r *remover) removeDoguServiceAccount(ctx context.Context, dogu *core.Dogu, serviceAccount core.ServiceAccount, senDoguCfg *config.DoguConfig) error {
 	logger := log.FromContext(ctx)
 	registryCredentialPath := "sa-" + serviceAccount.Type
-	doguConfig := r.registry.DoguConfig(dogu.GetSimpleName())
 
-	exists, err := serviceAccountExists(registryCredentialPath, doguConfig)
-	if err != nil {
-		return err
-	}
+	if exists := serviceAccountExists(registryCredentialPath, *senDoguCfg); !exists {
+		logger.Info(fmt.Sprintf("skipping removal of service account '%s' because it does not exist", registryCredentialPath))
 
-	if !exists {
-		logger.Info("skipping removal of service account because it does not exists")
 		return nil
 	}
 
@@ -104,7 +96,7 @@ func (r *remover) removeDoguServiceAccount(ctx context.Context, dogu *core.Dogu,
 		return nil
 	}
 
-	err = r.delete(ctx, serviceAccount, dogu, doguConfig, registryCredentialPath)
+	err = r.delete(ctx, serviceAccount, dogu, senDoguCfg, registryCredentialPath)
 	if err != nil {
 		return err
 	}
@@ -116,7 +108,7 @@ func (r *remover) delete(
 	ctx context.Context,
 	serviceAccount core.ServiceAccount,
 	dogu *core.Dogu,
-	doguConfig registry.ConfigurationContext,
+	senDoguCfg *config.DoguConfig,
 	registryCredentialPath string,
 ) error {
 	saDogu, err := r.doguFetcher.FetchInstalled(ctx, serviceAccount.Type)
@@ -134,9 +126,11 @@ func (r *remover) delete(
 		return fmt.Errorf("failed to execute service account remove command: %w", err)
 	}
 
-	err = doguConfig.DeleteRecursive(registryCredentialPath)
-	if err != nil {
-		return fmt.Errorf("failed to remove service account from config: %w", err)
+	updatedCfg := senDoguCfg.DeleteRecursive(config.Key(registryCredentialPath))
+	senDoguCfg.Config = updatedCfg
+
+	if lErr := writeConfig(ctx, senDoguCfg, r.sensitiveDoguRepo); lErr != nil {
+		return fmt.Errorf("failed write config for dogu %s after updating: %w", senDoguCfg.DoguName, lErr)
 	}
 
 	return nil
@@ -156,23 +150,6 @@ func (r *remover) executeCommand(ctx context.Context, consumerDogu *core.Dogu, s
 	_, err = r.executor.ExecCommandForPod(ctx, saPod, command, cloudogu.PodReady)
 	if err != nil {
 		return fmt.Errorf("failed to execute command: %w", err)
-	}
-
-	return nil
-}
-
-func (r *remover) removeCesControlServiceAccount(dogu *core.Dogu) error {
-	hostConfig := r.registry.HostConfig(k8sCesControl)
-	exists, err := hostConfig.Exists(dogu.GetSimpleName())
-	if err != nil {
-		return fmt.Errorf("failed to read host config for dogu %s", dogu.GetSimpleName())
-	}
-
-	if exists {
-		err = hostConfig.DeleteRecursive(dogu.GetSimpleName())
-		if err != nil {
-			return fmt.Errorf("failed to delete host config for dogu %s", dogu.GetSimpleName())
-		}
 	}
 
 	return nil