+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+**Breaking Change ahead**
+### Removed
+- [#184] Remove support for internal ETCD
+
+### Changed
+- [#184] Add k8s-registry lib in version 0.2.2 to use config maps for configuration instead of the etcd.
+  - This change requires all other installed dogus to use doguctl >= v0.12.1
 
 ## [v1.2.0] - 2024-06-12
 ### Added